tests: skip some SIDS hook tests if running tests as root
